fix#502 - When logged in, some pages have wrong navigation #502
fix#502 - When logged in, some pages have wrong navigation #502

This commit ends the modifications for all the issues about redirections
and logged in status. This and related previous commits includes the
followind redirections issues:
- Redirecting to the home recruiter page if this one tries to access
job-seeker pages, and vice versa.
- Redirecting to the loging page if the user is not logged in and try
to access a page what requires to be logged in.
- Redirecting to home user page when he tries to access / and is logged
in already
- Redirecting to /user/profile.create when the job-seeker has not
completed his profile.
- If a user is logged in and accesses the
about/term&conditions/howisworks pages, two aditional items menus are
show so he can see is logged in already.